Do not use an explicit owner and group
This enables the cookbook to run on windows machines where the administrator account is not named `Administrator`. Not using an explicit owner and group does not break the installation as it will just inherit the rights, which is actually the default behaviour you would expect.

Besides that it’s advised (for security reasons) to rename your administrator account, this poses a real problem on Azure.

Azure does not allow you to create a Windows VM with an account named `Administrator`. So you are required to choose a different name for the administrator account which prevents this cookbook from being able to run correctly.

Signed-off-by: Sander van Harmelen <sander@xanzy.io>