Replace broken perl install with a chocolatey perl install
StrawberryPerl from the Chocolatey project is a much better way to install perl on windows. This uses the build in chocolatey_package resource in Chef, but does require that chocolatey be installed first. Since that's outside the scope of this cookbook I've left it out, but it can easily be done with the chocolatey cookbook on Supermarket.

Signed-off-by: Tim Smith <tsmith@chef.io>